Rezolvare sarcina 2

+++ Controllers/BooksController.cs
@@ -22,42 +22,36 @@ namespace Apetrei_Alexandru_Lab2.Controllers
         // GET: Books
         public async Task<IActionResult> Index()
         {
-            var apetrei_Alexandru_Lab2Context = _context.Book.Include(b => b.Genre);
-            return View(await apetrei_Alexandru_Lab2Context.ToListAsync());
+            var books = _context.Book.Include(b => b.Genre).Include(b => b.Author);
+            return View(await books.ToListAsync());
         }
 
         // GET: Books/Details/5
         public async Task<IActionResult> Details(int? id)
         {
-            if (id == null)
-            {
-                return NotFound();
-            }
+            if (id == null) return NotFound();
 
             var book = await _context.Book
                 .Include(b => b.Genre)
+                .Include(b => b.Author)
                 .FirstOrDefaultAsync(m => m.ID == id);
-            if (book == null)
-            {
-                return NotFound();
-            }
 
+            if (book == null) return NotFound();
             return View(book);
         }
 
         // GET: Books/Create
         public IActionResult Create()
         {
-            ViewData["GenreID"] = new SelectList(_context.Set<Genre>(), "ID", "Name");
+            ViewData["GenreID"] = new SelectList(_context.Genre, "ID", "Name");
+            ViewData["AuthorID"] = new SelectList(_context.Author, "ID", "LastName");
             return View();
         }
 
         // POST: Books/Create
-        // To protect from overposting attacks, enable the specific properties you want to bind to.
-        // For more details, see http://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public async Task<IActionResult> Create([Bind("ID,Title,Author,Price,GenreID")] Book book)
+        public async Task<IActionResult> Create([Bind("ID,Title,Price,GenreID,AuthorID")] Book book)
         {
             if (ModelState.IsValid)
             {
@@ -65,38 +59,30 @@ namespace Apetrei_Alexandru_Lab2.Controllers
                 await _context.SaveChangesAsync();
                 return RedirectToAction(nameof(Index));
             }
-            ViewData["GenreID"] = new SelectList(_context.Set<Genre>(), "ID", "ID", book.GenreID);
+            ViewData["GenreID"] = new SelectList(_context.Genre, "ID", "Name", book.GenreID);
+            ViewData["AuthorID"] = new SelectList(_context.Author, "ID", "LastName", book.AuthorID);
             return View(book);
         }
 
         // GET: Books/Edit/5
         public async Task<IActionResult> Edit(int? id)
         {
-            if (id == null)
-            {
-                return NotFound();
-            }
+            if (id == null) return NotFound();
 
             var book = await _context.Book.FindAsync(id);
-            if (book == null)
-            {
-                return NotFound();
-            }
-            ViewData["GenreID"] = new SelectList(_context.Set<Genre>(), "ID", "Name", book.GenreID);
+            if (book == null) return NotFound();
+
+            ViewData["GenreID"] = new SelectList(_context.Genre, "ID", "Name", book.GenreID);
+            ViewData["AuthorID"] = new SelectList(_context.Author, "ID", "LastName", book.AuthorID);
             return View(book);
         }
 
         // POST: Books/Edit/5
-        // To protect from overposting attacks, enable the specific properties you want to bind to.
-        // For more details, see http://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public async Task<IActionResult> Edit(int id, [Bind("ID,Title,Author,Price,GenreID")] Book book)
+        public async Task<IActionResult> Edit(int id, [Bind("ID,Title,Price,GenreID,AuthorID")] Book book)
         {
-            if (id != book.ID)
-            {
-                return NotFound();
-            }
+            if (id != book.ID) return NotFound();
 
             if (ModelState.IsValid)
             {
@@ -107,37 +93,27 @@ namespace Apetrei_Alexandru_Lab2.Controllers
                 }
                 catch (DbUpdateConcurrencyException)
                 {
-                    if (!BookExists(book.ID))
-                    {
-                        return NotFound();
-                    }
-                    else
-                    {
-                        throw;
-                    }
+                    if (!_context.Book.Any(e => e.ID == book.ID)) return NotFound();
+                    else throw;
                 }
                 return RedirectToAction(nameof(Index));
             }
-            ViewData["GenreID"] = new SelectList(_context.Set<Genre>(), "ID", "Name", book.GenreID);
+            ViewData["GenreID"] = new SelectList(_context.Genre, "ID", "Name", book.GenreID);
+            ViewData["AuthorID"] = new SelectList(_context.Author, "ID", "LastName", book.AuthorID);
             return View(book);
         }
 
         // GET: Books/Delete/5
         public async Task<IActionResult> Delete(int? id)
         {
-            if (id == null)
-            {
-                return NotFound();
-            }
+            if (id == null) return NotFound();
 
             var book = await _context.Book
                 .Include(b => b.Genre)
+                .Include(b => b.Author)
                 .FirstOrDefaultAsync(m => m.ID == id);
-            if (book == null)
-            {
-                return NotFound();
-            }
 
+            if (book == null) return NotFound();
             return View(book);
         }
 
@@ -147,18 +123,9 @@ namespace Apetrei_Alexandru_Lab2.Controllers
         public async Task<IActionResult> DeleteConfirmed(int id)
         {
             var book = await _context.Book.FindAsync(id);
-            if (book != null)
-            {
-                _context.Book.Remove(book);
-            }
-
+            if (book != null) _context.Book.Remove(book);
             await _context.SaveChangesAsync();
             return RedirectToAction(nameof(Index));
-        }
-
-        private bool BookExists(int id)
-        {
-            return _context.Book.Any(e => e.ID == id);
         }
     }
 }
